+++ packages/detector/vite.cjs.config.ts
@@ -1,3 +1,4 @@
+import path from "node:path";
 // Copyright 2021-2025 Prosopo (UK) Ltd.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -11,13 +12,11 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-import path from "node:path";
 import { ViteCommonJSConfig } from "@prosopo/config";
 
 export default function () {
 	return ViteCommonJSConfig(
-		"detector",
-		path.resolve("./tsconfig.cjs.json"),
-		"src/index.js",
+		path.basename("."),
+		path.resolve("./tsconfig.json"),
 	);
 }