+++ packages/provider/src/api/ignoreMiddleware.ts
@@ -11,11 +11,16 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-import { ApiPrefix } from "@prosopo/types";
+import { ApiPrefix, PublicApiPaths } from "@prosopo/types";
 import type { NextFunction, Request, Response } from "express";
 
 export function ignoreMiddleware() {
 	return (req: Request, res: Response, next: NextFunction) => {
+		if (req.originalUrl.indexOf(PublicApiPaths.Healthz) !== -1) {
+			// If the request is for a health endpoint, we allow it to pass through
+			return next();
+		}
+
 		// Ignore non-api routes
 		if (req.originalUrl.indexOf(ApiPrefix) === -1) {
 			res.statusCode = 404;