+++ packages/user-access-policy/src/index.ts
@@ -13,62 +13,82 @@
 // limitations under the License.
 
 import type { ApiRoutesProvider } from "@prosopo/api-route";
-import type { Logger } from "@prosopo/common";
-import type { Model } from "mongoose";
-import type { BlacklistInspector } from "./blacklistInspector.js";
-import { apiRulePaths } from "./rules/api/apiRulePaths.js";
-import { ApiRuleRoutesProvider } from "./rules/api/apiRuleRoutesProvider.js";
-import { getExpressApiRuleRateLimits } from "./rules/api/getExpressApiRuleRateLimits.js";
-import type {
-	ApiInsertManyRulesArgsOutputSchema,
-	ApiInsertManyRulesArgsSchema,
-} from "./rules/api/insertMany/apiInsertManyRulesArgsSchema.js";
-import { BlacklistRulesInspector } from "./rules/blacklistRulesInspector.js";
-import { ImageCaptchaConfigRulesResolver } from "./rules/imageCaptchaConfigRulesResolver.js";
-import { RulesMongooseStorage } from "./rules/mongoose/rulesMongooseStorage.js";
-import { getRuleMongooseSchema } from "./rules/mongoose/schemas/getRuleMongooseSchema.js";
-import type { Rule } from "./rules/rule/rule.js";
-import type { RulesStorage } from "./rules/storage/rulesStorage.js";
+import {
+	type AccessPolicy,
+	AccessPolicyType,
+	type AccessRuleExtended,
+	type PolicyScope,
+	type UserScope,
+	type UserScopeApiInput,
+	type UserScopeApiOutput,
+	accessPolicySchema,
+	accessRuleSchemaExtended,
+	policyScopeSchema,
+} from "#policy/accessPolicy.js";
+import { type PolicyFilter, ScopeMatch } from "#policy/accessPolicyResolver.js";
+import {
+	type AccessRule,
+	type AccessRulesStorage,
+	accessRuleSchema,
+} from "#policy/accessRules.js";
+import {
+	AccessRuleApiRoutes,
+	accessRuleApiPaths,
+	getExpressApiRuleRateLimits,
+} from "#policy/api/accessRuleApiRoutes.js";
+import {
+	type DeleteAllRulesEndpointSchema,
+	deleteAllRulesEndpointSchema,
+} from "#policy/api/deleteAllRulesEndpoint.js";
+import {
+	type DeleteRulesEndpointSchemaInput,
+	type DeleteRulesEndpointSchemaOutput,
+	deleteRulesEndpointSchema,
+} from "#policy/api/deleteRulesEndpoint.js";
+import {
+	type InsertManyRulesEndpointInputSchema,
+	type InsertManyRulesEndpointOutputSchema,
+	insertRulesEndpointSchema,
+} from "#policy/api/insertRulesEndpoint.js";
+import { createRedisAccessRulesStorage } from "#policy/redis/redisAccessRules.js";
+import { createRedisAccessRulesIndex } from "#policy/redis/redisAccessRulesIndex.js";
+import { userScopeInputSchema } from "./accessPolicy.js";
 
-const createBlacklistInspector = (
-	rulesStorage: RulesStorage,
-	logger: Logger,
-): BlacklistInspector => {
-	return new BlacklistRulesInspector(rulesStorage, logger);
-};
-
-const createImageCaptchaConfigResolver = (
-	rulesStorage: RulesStorage,
-	logger: Logger,
-): ImageCaptchaConfigRulesResolver => {
-	return new ImageCaptchaConfigRulesResolver(rulesStorage, logger);
-};
-
-const createApiRuleRoutesProvider = (
-	rulesStorage: RulesStorage,
+export const createApiRuleRoutesProvider = (
+	rulesStorage: AccessRulesStorage,
 ): ApiRoutesProvider => {
-	return new ApiRuleRoutesProvider(rulesStorage);
-};
-
-const createMongooseRulesStorage = (
-	logger: Logger,
-	readingModel: Model<Rule> | null,
-	writingModel: Model<Rule> | null = null,
-): RulesStorage => {
-	return new RulesMongooseStorage(logger, readingModel, writingModel);
+	return new AccessRuleApiRoutes(rulesStorage);
 };
 
 export {
-	type Rule,
-	type RulesStorage,
-	type BlacklistInspector,
-	type ApiInsertManyRulesArgsSchema,
-	type ApiInsertManyRulesArgsOutputSchema,
-	createMongooseRulesStorage,
-	createImageCaptchaConfigResolver,
-	createBlacklistInspector,
-	createApiRuleRoutesProvider,
-	getRuleMongooseSchema,
+	type AccessPolicy,
+	type PolicyScope,
+	type AccessRulesStorage,
+	type PolicyFilter,
+	type DeleteRulesEndpointSchemaOutput,
+	type DeleteRulesEndpointSchemaInput,
+	type DeleteAllRulesEndpointSchema,
+	type InsertManyRulesEndpointInputSchema,
+	type InsertManyRulesEndpointOutputSchema,
+	type AccessRule,
+	type UserScope,
+	type UserScopeApiInput,
+	type UserScopeApiOutput,
+	type AccessRuleExtended,
+	AccessPolicyType,
+	ScopeMatch,
+	// redis
+	createRedisAccessRulesIndex,
+	createRedisAccessRulesStorage,
+	// api
+	accessRuleApiPaths,
+	accessRuleSchema,
+	accessRuleSchemaExtended,
+	accessPolicySchema,
+	policyScopeSchema,
+	insertRulesEndpointSchema,
+	deleteAllRulesEndpointSchema,
+	deleteRulesEndpointSchema,
 	getExpressApiRuleRateLimits,
-	apiRulePaths,
+	userScopeInputSchema,
 };