+++ packages/util-crypto/src/nacl/index.ts
@@ -1,0 +1,8 @@
+// Copyright 2017-2025 @polkadot/util-crypto authors & contributors
+// SPDX-License-Identifier: Apache-2.0
+
+/**
+ * @summary Implements [NaCl](http://nacl.cr.yp.to/) secret-key authenticated encryption, public-key authenticated encryption
+ */
+export { naclDecrypt } from "./decrypt.js";
+export { naclEncrypt } from "./encrypt.js";