+++ demos/cypress-shared/CHANGELOG.md
@@ -1,5 +1,171 @@
 # @prosopo/cypress-shared
 
+## 2.8.14
+### Patch Changes
+
+- 1f3a02f: Release 3.1.8
+
+## 2.8.13
+### Patch Changes
+
+- Updated dependencies [44ffda2]
+- Updated dependencies [a49b538]
+  - @prosopo/config@3.1.4
+  - @prosopo/types@3.0.7
+  - @prosopo/util@3.0.6
+
+## 2.8.12
+### Patch Changes
+
+- 828066d: remove empty test npm scripts, add missing npm test scripts
+- 91bbe87: configure typecheck before bundle for vue packages
+- 91bbe87: make typecheck script always recompile
+- 346e092: NODE_ENV default to "development"
+- 5d36e05: remove tsc --force
+- Updated dependencies [828066d]
+- Updated dependencies [df4e030]
+- Updated dependencies [91bbe87]
+- Updated dependencies [3ef4fd2]
+- Updated dependencies [91bbe87]
+- Updated dependencies [346e092]
+- Updated dependencies [5d36e05]
+  - @prosopo/types@3.0.6
+  - @prosopo/config@3.1.3
+  - @prosopo/util@3.0.5
+
+## 2.8.11
+### Patch Changes
+
+- eb71691: configure typecheck before bundle for vue packages
+- eb71691: make typecheck script always recompile
+- Updated dependencies [eb71691]
+- Updated dependencies [eb71691]
+  - @prosopo/types@3.0.5
+  - @prosopo/util@3.0.4
+  - @prosopo/config@3.1.2
+
+## 2.8.10
+### Patch Changes
+
+- 3573f0b: fix npm scripts bundle command
+- 3573f0b: build using vite, typecheck using tsc
+- efd8102: Add tests for unwrap error helper
+- 3573f0b: standardise all vite based npm scripts for bundling
+- Updated dependencies [52dbf21]
+- Updated dependencies [93d5e50]
+- Updated dependencies [3573f0b]
+- Updated dependencies [3573f0b]
+- Updated dependencies [efd8102]
+- Updated dependencies [93d5e50]
+- Updated dependencies [63519d7]
+- Updated dependencies [3573f0b]
+- Updated dependencies [2d0dd8a]
+  - @prosopo/util@3.0.3
+  - @prosopo/types@3.0.4
+  - @prosopo/config@3.1.1
+
+## 2.8.9
+### Patch Changes
+
+- Updated dependencies [b0d7207]
+  - @prosopo/types@3.0.3
+
+## 2.8.8
+### Patch Changes
+
+  - @prosopo/util@3.0.2
+
+## 2.8.7
+### Patch Changes
+
+  - @prosopo/util@3.0.1
+
+## 2.8.6
+### Patch Changes
+
+- Updated dependencies [f682f0c]
+  - @prosopo/types@3.0.2
+
+## 2.8.5
+### Patch Changes
+
+  - @prosopo/types@3.0.1
+
+## 2.8.4
+### Patch Changes
+
+- Updated dependencies [64b5bcd]
+  - @prosopo/types@3.0.0
+  - @prosopo/util@3.0.0
+
+## 2.8.3
+### Patch Changes
+
+- Updated dependencies [aee3efe]
+  - @prosopo/types@2.10.0
+
+## 2.8.2
+### Patch Changes
+
+- 86c22b8: structured logging
+- Updated dependencies [86c22b8]
+  - @prosopo/types@2.9.1
+  - @prosopo/util@2.6.1
+
+## 2.8.1
+### Patch Changes
+
+- Updated dependencies [30bb383]
+  - @prosopo/types@2.9.0
+
+## 2.8.0
+### Minor Changes
+
+- 8f0644a: Taking required functions from polkadot/keyring and polkadot/util-crypto in-house and removing WASM dependencies. Adding @scure JS-based sr25519 function instead.
+
+### Patch Changes
+
+- Updated dependencies [8f0644a]
+  - @prosopo/types@2.8.0
+
+## 2.7.4
+
+### Patch Changes
+
+- @prosopo/types@2.7.1
+
+## 2.7.3
+
+### Patch Changes
+
+- 5656b0c: Adding cypress tests for invisible
+
+## 2.7.2
+
+### Patch Changes
+
+- Updated dependencies [6e1aef6]
+  - @prosopo/types@2.7.0
+
+## 2.7.1
+
+### Patch Changes
+
+- Updated dependencies [6ff193a]
+  - @prosopo/types@2.6.2
+
+## 2.7.0
+
+### Minor Changes
+
+- cae05fd: Turning async defer on for cypress test suite
+
+### Patch Changes
+
+- 52feffc: Adjustable difficulty img captcha
+- Updated dependencies [52feffc]
+  - @prosopo/types@2.6.1
+
 ## 2.6.0
 
 ### Minor Changes